Add navbar to main page
